prodcuts addition endpoint completed

+++ lib/features/admin/screens/add_product_screen.dart
@@ -51,7 +51,7 @@ class _AddProductScreenState extends State<AddProductScreen> {
           name: productNameController.text,
           description: descriptionController.text,
           price: double.parse(priceController.text),
-          quantity: double.parse(quantityController.text),
+          quantity: int.parse(quantityController.text),
           category: category,
           images: images);
     }
